Refactor OAuth module name
