+++ Apps/OAuth/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class OAuthConfig(AppConfig):
+    name = "OAuth"
