+++ Apps/OAuth/tests/test_google_view.py
@@ -23,7 +23,7 @@ class TestGoogleAuthView:
     def test_url(self) -> None:
         assert self.url() == "/api/oauth/google/"
 
-    @patch("SocialAuth.serializers.verify_oauth2_token")
+    @patch("OAuth.serializers.verify_oauth2_token")
     def test_google_view_creates_new_user(
         self, mock_verify_oauth2_token: MagicMock, client: APIClient
     ) -> None:
@@ -39,7 +39,7 @@ class TestGoogleAuthView:
         assert response.status_code == 200
         assert User.objects.filter(email=email).exists()
 
-    @patch("SocialAuth.serializers.verify_oauth2_token")
+    @patch("OAuth.serializers.verify_oauth2_token")
     def test_google_view_returns_user_login_data(
         self, mock_verify_oauth2_token: MagicMock, client: APIClient
     ) -> None:
