+++ Apps/OAuth/tests/test_twitter_serializer.py
@@ -4,7 +4,7 @@ from mock import patch
 from mock.mock import MagicMock
 from rest_framework.serializers import ValidationError
 
-from SocialAuth.serializers import TwitterOAuthSerializer
+from OAuth.serializers import TwitterOAuthSerializer
 from Users.models import User
 from Users.serializers import UserAuthSerializer
