+++ Apps/OAuth/views.py
@@ -4,9 +4,9 @@ from rest_framework.response import Response
 from rest_framework.serializers import Serializer
 from rest_framework.status import HTTP_200_OK as OK
 
-from SocialAuth.serializers import FacebookOAuthSerializer
-from SocialAuth.serializers import GoogleOAuthSerializer
-from SocialAuth.serializers import TwitterOAuthSerializer
+from OAuth.serializers import FacebookOAuthSerializer
+from OAuth.serializers import GoogleOAuthSerializer
+from OAuth.serializers import TwitterOAuthSerializer
 
 
 class GenericOAuthView(GenericAPIView):
@@ -19,13 +19,13 @@ class GenericOAuthView(GenericAPIView):
         return Response(serializer.data, status=OK)
 
 
-class GoogleSocialAuthView(GenericOAuthView):
+class GoogleOAuthView(GenericOAuthView):
     serializer_class: GoogleOAuthSerializer = GoogleOAuthSerializer
 
 
-class FacebookSocialAuthView(GenericOAuthView):
+class FacebookOAuthView(GenericOAuthView):
     serializer_class: FacebookOAuthSerializer = FacebookOAuthSerializer
 
 
-class TwitterSocialAuthView(GenericOAuthView):
+class TwitterOAuthView(GenericOAuthView):
     serializer_class: TwitterOAuthSerializer = TwitterOAuthSerializer
