+++ Project/urls.py
@@ -22,7 +22,7 @@ urlpatterns: list = [
     path("admin/", admin.site.urls),
     path("api/", include(("Users.urls", "users"), namespace="users")),
     path("api/", include(("Emails.urls", "emails"), namespace="emails")),
-    path("api/oauth/", include(("SocialAuth.urls", "auth"), namespace="auth")),
+    path("api/oauth/", include(("OAuth.urls", "auth"), namespace="oauth")),
     path(
         "api/password_reset/",
         include("django_rest_passwordreset.urls", namespace="password_reset"),
